[Code Refactoring] Seprate out getCommonEditTextFieldProps method into new file as a part of code Refactoring
#5
